+++ dither_go/__init__.py
@@ -1,7 +1,7 @@
 # Copyright 2023, tfuxu <https://github.com/tfuxu>
 # SPDX-License-Identifier: GPL-3.0-or-later
 
-__version__ = "0.2.1"
+__version__ = "0.3.0"
 
 from .wrapper import *
 from .matrices import *